server-side calls: rename cmk/base/plugins/config_generation
