+++ cmk/base/sources/_builder.py
@@ -39,7 +39,7 @@ import cmk.base.config as config
 from cmk.base.api.agent_based.register.snmp_plugin_store import make_plugin_store
 from cmk.base.config import ConfigCache
 from cmk.base.ip_lookup import AddressFamily
-from cmk.base.plugins.config_generation import load_special_agents
+from cmk.base.plugins.server_side_calls import load_special_agents
 from cmk.base.server_side_calls import SpecialAgent
 
 from ._api import Source
